# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bijou AI WhatsApp Enterprise",
    "version": "2.2.0",
    "description": "Enterprise-grade AI customer support for WhatsApp with multi-language support, human escalation, and Google Sheets integration. Built for property agents and SMEs.",
    "contact": {
      "name": "Bijou AI Support Team",
      "email": "support@mybijou.xyz",
      "url": "https://app.mybijou.xyz/static/help.html"
    },
    "license": {
      "name": "Commercial License",
      "url": "https://mybijou.xyz/terms"
    }
  },
  "paths": {
...
    "/api/settings/testing-mode": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Testing Mode",
        "description": "Toggle testing mode and set test numbers.\n\nWhen testing mode is enabled, Bijou will ONLY reply to numbers in test_numbers list.\nThis protects production users during testing.\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nBody:\n    testing_mode: true/false\n    test_numbers: Array of phone numbers (e.g., [\"+60143856929\"])\n\nReturns:\n    TestingModeResponse with updated settings",
        "operationId": "update_testing_mode_api_settings_testing_mode_put",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestingModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestingModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/ignore-list": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Ignore List",
        "description": "Update ignore/private number lists.\n\nNumbers in these lists will NEVER receive auto-replies from Bijou.\nUse this to exclude owner's personal number, test numbers, or VIP contacts.\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nBody:\n    ignore_numbers: Array of numbers to ignore\n    private_numbers: Array of private numbers (alias for ignore)\n\nReturns:\n    IgnoreListResponse with updated lists",
        "operationId": "update_ignore_list_api_settings_ignore_list_put",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IgnoreListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IgnoreListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/business-hours": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Business Hours",
        "description": "Update business hours configuration.\n\nWhen enabled, Bijou will only auto-reply during specified business hours.\nOutside hours, a custom out-of-hours message is sent (if provided).\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nBody:\n    enabled: true/false\n    timezone: Timezone string (e.g., \"Asia/Kuala_Lumpur\")\n    schedule: Object with daily schedules\n    out_of_hours_message: Custom message for outside hours\n\nReturns:\n    BusinessHoursResponse with updated configuration",
        "operationId": "update_business_hours_api_settings_business_hours_put",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessHoursRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessHoursResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/auto-reply": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Auto Reply",
        "description": "Toggle auto-reply and set welcome message.\n\nWhen auto-reply is disabled, Bijou will NOT respond to any messages.\nUse this to temporarily pause the AI assistant.\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nBody:\n    auto_reply_enabled: true/false\n    welcome_message: Optional custom welcome message\n\nReturns:\n    AutoReplyResponse with updated settings",
        "operationId": "update_auto_reply_api_settings_auto_reply_put",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoReplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoReplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoReplyRequest": {
        "properties": {
          "auto_reply_enabled": {
            "type": "boolean",
            "title": "Auto Reply Enabled"
          },
          "welcome_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Welcome Message"
          }
        },
        "type": "object",
        "required": [
          "auto_reply_enabled"
        ],
        "title": "AutoReplyRequest",
        "description": "Toggle auto-reply"
      },
      "AutoReplyResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "auto_reply_enabled": {
            "type": "boolean",
            "title": "Auto Reply Enabled"
          },
          "welcome_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Welcome Message"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "auto_reply_enabled",
          "welcome_message",
          "message"
        ],
        "title": "AutoReplyResponse",
        "description": "Response after updating auto-reply"
      },
...
      "BusinessHoursRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Kuala_Lumpur"
          },
          "schedule": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BusinessHoursSchedule"
            },
            "type": "object",
            "title": "Schedule"
          },
          "out_of_hours_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Of Hours Message"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "schedule"
        ],
        "title": "BusinessHoursRequest",
        "description": "Update business hours configuration"
      },
      "BusinessHoursResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "business_hours": {
            "additionalProperties": true,
            "type": "object",
            "title": "Business Hours"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "business_hours",
          "message"
        ],
        "title": "BusinessHoursResponse",
        "description": "Response after updating business hours"
      },
      "BusinessHoursSchedule": {
        "properties": {
          "start": {
            "type": "string",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "title": "End"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end",
          "enabled"
        ],
        "title": "BusinessHoursSchedule",
        "description": "Daily schedule"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IgnoreListRequest": {
        "properties": {
          "ignore_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ignore Numbers",
            "default": []
          },
          "private_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Private Numbers",
            "default": []
          }
        },
        "type": "object",
        "title": "IgnoreListRequest",
        "description": "Update ignore/private number list"
      },
      "IgnoreListResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "ignore_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ignore Numbers"
          },
          "private_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Private Numbers"
          },
          "total_ignored": {
            "type": "integer",
            "title": "Total Ignored"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "ignore_numbers",
          "private_numbers",
          "total_ignored",
          "message"
        ],
        "title": "IgnoreListResponse",
        "description": "Response after updating ignore list"
      },
...
      "TestingModeRequest": {
        "properties": {
          "testing_mode": {
            "type": "boolean",
            "title": "Testing Mode"
          },
          "test_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Test Numbers",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "testing_mode"
        ],
        "title": "TestingModeRequest",
        "description": "Toggle testing mode and set test numbers"
      },
      "TestingModeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "testing_mode": {
            "type": "boolean",
            "title": "Testing Mode"
          },
          "test_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Test Numbers"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "testing_mode",
          "test_numbers",
          "message"
        ],
        "title": "TestingModeResponse",
        "description": "Response after updating testing mode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}